langcache-operations/embeddings/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
requests==2.31.0
flask==2.0.1
werkzeug==2.0.3
numpy==1.24.3
scipy==1.10.1
scikit-learn==1.3.0
huggingface-hub==0.20.1
transformers==4.36.2
sentence-transformers==2.2.2
torch==2.0.1
pillow==10.0.0
